Use updated kafka lib with indefinite offset retention.
Pessimistically use kafka.OffsetNewest as the initial offset until we
are confident that the issue is entirely fixed. The worst case scenario
using OffsetNewest is lost messages. For the import process this means
trying the import again. The worst case when using OffsetOldest is
replaying messages and making data inconsistent. We should consider
changing all services to use kafka.OffsetOldest when the issue is fixed,
so that no messages are lost when the service first starts.